Migrate Site Title Example block to a separate component. Fix whitespace issue.
