+++ assets/stylesheets/_components.scss
@@ -115,6 +115,7 @@
 @import 'components/site-icon/style';
 @import 'components/site-selector/style';
 @import 'components/site-selector-modal/style';
+@import 'components/site-title-example/style';
 @import 'components/sites-dropdown/style';
 @import 'components/sites-popover/style';
 @import 'components/social-logo/style';
