+++ client/signup/steps/site-title/style.scss
@@ -21,62 +21,7 @@
 
 	.signup-site-title__button {
 		position: absolute;
-			top: 7px;
-			right: 8px;
-	}
-}
-
-.example-site-title {
-	box-sizing: border-box;
-	position: relative;
-	padding-top: 42px;
-	padding-right: 348px;
-	max-width: 600px;
-	min-height: 120px;
-	margin: 30px auto 60px auto;
-
-	p {
-		color: darken( $gray, 20% );
-		text-align: right;
-		margin-bottom: 8px;
-	}
-
-	a {
-		text-decoration: underline;
-	}
-
-	@include breakpoint( "<660px" ) {
-		max-width: 100%;
-		padding: 32px 32px 0;
-		margin: 0 auto;
-
-		p {
-			text-align: center;
-			margin-bottom: 16px;
-		}
-	}
-}
-
-.example-site-title__image {
-	pointer-events: none;
-	position: absolute;
-	top: 0;
-	right: 0;
-	width: 300px;
-
-	&:before {
-		display: block;
-		content: '';
-		position: absolute;
-		right: 0;
-		bottom: 0;
-		left: 0;
-		height: 30px;
-		background: linear-gradient( to bottom, rgba( $gray-light, 0 ) 0%, rgba( $gray-light, 1 ) 100% );
-	}
-
-	@include breakpoint( "<660px" ) {
-		position: relative;
-		margin: 0 auto;
+		top: 7px;
+		right: 8px;
 	}
 }
